refactor: extract ipc registry, remote urls and workspace file resolvers
