+++ packages/contracts/package.json
@@ -12,6 +12,7 @@
     "./invite-links": "./src/invite-links.ts",
     "./ipc": "./src/ipc.ts",
     "./mobile-connect": "./src/mobile-connect.ts",
+    "./remote-urls": "./src/remote-urls.ts",
     "./runtime-values": "./src/runtime-values.ts",
     "./team-protocol": "./src/team-protocol/index.ts",
     "./team-protocol/current": "./src/team-protocol/current.ts",
