+++ src/main/remote-server-manager.ts
@@ -74,6 +74,7 @@ import {
   isSidebarLayoutSnapshot,
   isTeamRealtimeEvent,
 } from "@openbot/contracts/ipc";
+import { remoteAgentAvatarUrl, remoteAttachmentPreviewUrl, remoteServerLogoUrl } from "@openbot/contracts/remote-urls";
 import {
   type DynamicRecord,
   isBoolean,
@@ -119,6 +120,11 @@ import {
 } from "./team-webrtc-client-transport";
 
 export { isValidRemoteApiUrl } from "@openbot/contracts/invite-links";
+export {
+  remoteAgentAvatarUrl,
+  remoteAttachmentPreviewUrl,
+  remoteServerLogoUrl,
+} from "@openbot/contracts/remote-urls";
 
 interface StoredRemoteServer {
   id: string;
@@ -2859,21 +2865,4 @@ function addRemotePreviewUrls<T>(value: T, serverId: string): T {
   }
   for (const item of Object.values(record)) addRemotePreviewUrls(item, serverId);
   return value;
-}
-
-export function remoteAttachmentPreviewUrl(serverId: string, attachmentId: string): string {
-  return `openbot-remote-attachment://${encodeURIComponent(serverId)}/${encodeURIComponent(attachmentId)}`;
-}
-
-export function remoteAgentAvatarUrl(serverId: string, botId: string, sourceUrl: string): string {
-  const source = new URL(sourceUrl);
-  const target = new URL(`openbot-remote-avatar://${encodeURIComponent(serverId)}/${encodeURIComponent(botId)}`);
-  target.search = source.search;
-  return target.toString();
-}
-
-export function remoteServerLogoUrl(serverId: string, version: string): string {
-  const target = new URL(`openbot-remote-server-logo://${encodeURIComponent(serverId)}/logo`);
-  target.searchParams.set("v", version);
-  return target.toString();
 }
